refactor: move `ImmutableRequest` creation to `expo-server`
